Rename gym map view to just gym map
